Move getCaveats to a separate helper module
