+++ test/unit/src/common/actions/t_auth-store.js
@@ -28,7 +28,6 @@ const {
   extractExpiresCaveat,
   extractSSOCaveat,
   getAccountInfo,
-  getCaveats,
   getSSODischarge,
   signAgreementSuccess,
   signIntoStore,
@@ -52,37 +51,6 @@ describe('store authentication actions', () => {
 
   beforeEach(() => {
     store = mockStore(initialState);
-  });
-
-  context('getCaveats', () => {
-    it('generates a sequence of caveats', () => {
-      const macaroon = new MacaroonsBuilder('location', 'key', 'id')
-        .add_first_party_caveat('caveat 1')
-        .add_third_party_caveat('external', 'external key', 'caveat 2')
-        .add_first_party_caveat('caveat 3')
-        .getMacaroon();
-      const caveats = [];
-      for (const caveat of getCaveats(macaroon)) {
-        caveats.push(caveat);
-      }
-      expect(caveats).toMatch([
-        {
-          caveatId: 'caveat 1',
-          verificationKeyId: '',
-          location: ''
-        },
-        {
-          caveatId: 'caveat 2',
-          verificationKeyId: /./,
-          location: 'external'
-        },
-        {
-          caveatId: 'caveat 3',
-          verificationKeyId: '',
-          location: ''
-        }
-      ]);
-    });
   });
 
   context('extractExpiresCaveat', () => {
